Move thermal angle-energy distributions into separate file
